requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0

# Web crawling
scrapy==2.11.2
beautifulsoup4==4.12.3
lxml==5.2.1
requests==2.31.0

# PDF / Document processing
pypdf2==3.0.1
python-magic==0.4.27

# Azure SDKs
azure-identity==1.16.0
azure-storage-blob==12.20.0
azure-search-documents==11.6.0
azure-ai-documentintelligence==1.0.0
openai==1.30.1

# Text processing / chunking
tiktoken==0.7.0
langchain-text-splitters==0.2.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
